FlaskBase/requirements.txt
Flask==2.1.3
Flask-WTF
SQLAlchemy==1.4
Flask-Migrate
Flask-Login
flask_restful
pandas
openpyxl
werkzeug==2.0
Pillow == 2.1.0